refactor: 인기순 대량 작업 레포지토리 추가
